Added a built-in unary procedure mag to return the magnitude (absolute value) of a Complex as a Number. It is a semantic error of the argument is not Complex.
Added a built-in unary procedure arg to return the argument (angle or phase) of a Complex as a Number in radians. It is a semantic error of the argument is not Complex.

Added a built-in unary procedure conj to return the conjugate of a Complex argument. It is a semantic error of the argument is not Complex.