Extended the procedure sqrt from task 1 to also return the square root of a Complex argument or a negative Number argument
Extended the procedure ^ from task 1 to return the exponential of a to the power b for any combination of Number and Complex types on a and b. The result is of type Number only when both arguments are of type Number.

Fixed a bug with multiplication of two complex numbers.